Changes
Jackson bidirection anti recursion annotations, further changes added. (@JsonBackReference and @JsonManagedReference). Changes made in the models.

 findAll, findById and find by composite ids added to each repository and associated controller.

 Added searches for retrieving StockItems by manufacturer, itemcategory or 'like'. Note that 'like' doesn't work at this point.

+++ Paddy30REST/src/main/java/org/fyp/repository/ItemCategoryRepository.java
@@ -8,6 +8,6 @@ import javax.transaction.Transactional;
 
 @RepositoryRestResource(collectionResourceRel = "itemCategory", path = "ItemCategories")
 @Transactional
-public interface ItemCategoryRepository extends JpaRepository<ItemCategory, Long> {
-    ItemCategory findByItemCategoryId(Long itemCategoryId);
+public interface ItemCategoryRepository extends JpaRepository<ItemCategory, Integer> {
+    ItemCategory findByItemCategoryId(Integer itemCategoryId);
 }
